% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_init_exp.R
\name{center_views}
\alias{center_views}
\title{Center view-data}
\usage{
center_views(pb_dat_list)
}
\arguments{
\item{pb_dat_list}{List of SummarizedExperiment generated from \code{MOFAcellulaR::filt_profiles()}}
}
\value{
A named list of SummarizedExperiments per cell type provided with centered pseudobulk profiles
}
\description{
Centers each element of a list of SummarizedExperiments
}
\details{
Given that the MOFA model in general uses centered data, when interested in projecting new
data to a new manifold, it is needed to perform centering.
}
\examples{
inputs_dir <- base::system.file("extdata", package = "MOFAcellulaR")
load(file.path(inputs_dir, "testpbcounts.rda"))
load(file.path(inputs_dir, "testcoldata.rda"))
pb_obj <- create_init_exp(counts = testpbcounts,  coldata = testcoldata)

ct_list <- filt_profiles(pb_dat = pb_obj,
                         cts = c("Fib","CM"),
                         ncells = 5,
                         counts_col = "cell_counts",
                         ct_col = "cell_type")

ct_list <- filt_gex_byexpr(pb_dat_list = ct_list,
                           min.count = 5,
                           min.prop = 0.25)

ct_list <- tmm_trns(pb_dat_list = ct_list,
                    scale_factor = 1000000)

ct_list <- center_views(pb_dat_list = ct_list)

multiview_dat <- pb_dat2MOFA(pb_dat_list = ct_list)
}
