% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_init_exp.R
\name{create_init_exp}
\alias{create_init_exp}
\title{Single-cell SummarizedExperiment object}
\usage{
create_init_exp(counts, coldata)
}
\arguments{
\item{counts}{Named numeric matrix with features in rows and samples in columns.}

\item{coldata}{A data frame containing the annotations of the samples.}
}
\value{
SummarizedExperiment with provided data
}
\description{
Creates a SummarizedExperiment object necessary to build a multi-view representation.
}
\details{
This function is the first step for a multicellular factor analysis.
It collects in a single object the pseudobulk counts of a single cell experiment
and its annotations.
}
\examples{
x <- matrix(rep(1,4),nrow = 2, ncol = 2)
rownames(x) <- c("featA", "featB")
colnames(x) <- c("Sample1", "Sample2")
y <- data.frame(cell_type = c("A","A"),
donor_id = c("Sample1", "Sample2"),
cell_counts = c(30,30))
se <- create_init_exp(x, y)
}
