% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_basic.R
\name{filt_gex_byexpr}
\alias{filt_gex_byexpr}
\title{Filter lowly expressed genes from pseudobulk profiles}
\usage{
filt_gex_byexpr(pb_dat_list, min.count, min.prop)
}
\arguments{
\item{pb_dat_list}{List of SummarizedExperiment generated from \code{MOFAcellulaR::filt_profiles()}}

\item{min.count}{Numeric, minimum counts per sample to be considered. Check \code{?edgeR::filterByExpr()} for details.}

\item{min.prop}{Numeric, minimum proportion of samples containing the minimum counts. Check \code{?edgeR::filterByExpr()} for details.}
}
\value{
A named list of SummarizedExperiments per cell type provided with filtered pseudobulk profiles
}
\description{
Filter lowly expressed genes from pseudobulk profiles using \code{edgeR}.
}
\details{
This function wraps \code{edgeR::filterByExpr()} to be applied to lists of SummarizedExperiments.It assumes
that all samples are part of the same group.
}
\examples{
inputs_dir <- base::system.file("extdata", package = "MOFAcellulaR")
load(file.path(inputs_dir, "testpbcounts.rda"))
load(file.path(inputs_dir, "testcoldata.rda"))

pb_obj <- create_init_exp(counts = testpbcounts,
                          coldata = testcoldata)

ct_list <- filt_profiles(pb_dat = pb_obj,
                         cts = c("Fib","CM"),
                         ncells = 5,
                         counts_col = "cell_counts",
                         ct_col = "cell_type")

ct_list <- filt_gex_byexpr(pb_dat_list = ct_list,
                           min.count = 5,
                           min.prop = 0.25)
}
