% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_basic.R
\name{filt_samples_bycov}
\alias{filt_samples_bycov}
\title{Filter samples that have a limited coverage of features in a view}
\usage{
filt_samples_bycov(pb_dat_list, prop_coverage = 0.9)
}
\arguments{
\item{pb_dat_list}{List of SummarizedExperiment generated from \code{MOFAcellulaR::filt_profiles()}}

\item{prop_coverage}{Numeric, minimum proportion of coverage of features (different from 0).}
}
\value{
A named list of SummarizedExperiments per cell type provided with filtered pseudobulk profiles
}
\description{
Excludes samples from views with a coverage lower than the one stated
}
\details{
This function allows the user to exclude samples that have a limited number of features that could create potential outliers
}
\examples{
inputs_dir <- base::system.file("extdata", package = "MOFAcellulaR")
load(file.path(inputs_dir, "testpbcounts.rda"))
load(file.path(inputs_dir, "testcoldata.rda"))

pb_obj <- create_init_exp(counts = testpbcounts,
                          coldata = testcoldata)

ct_list <- filt_profiles(pb_dat = pb_obj,
                         cts = c("Fib","CM"),
                         ncells = 5,
                         counts_col = "cell_counts",
                         ct_col = "cell_type")

ct_list <- filt_gex_byexpr(pb_dat_list = ct_list,
                           min.count = 5,
                           min.prop = 0.25)

ct_list <- filt_views_bygenes(pb_dat_list = ct_list,
                              ngenes = 15)

ct_list <- filt_samples_bycov(pb_dat_list = ct_list,
                              prop_coverage = 0.9)


}
