% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_basic.R
\name{filt_views_bysamples}
\alias{filt_views_bysamples}
\title{Filter views with not enough samples}
\usage{
filt_views_bysamples(pb_dat_list, nsamples)
}
\arguments{
\item{pb_dat_list}{List of SummarizedExperiment generated from \code{MOFAcellulaR::filt_profiles()}}

\item{nsamples}{Numeric, minimum number of samples per view.}
}
\value{
A named list of SummarizedExperiments per cell type provided with filtered pseudobulk profiles
}
\description{
Filter complete views based on their number of profiled samples
}
\details{
This function allows the user to control the number of minimum samples per view
}
\examples{
inputs_dir <- base::system.file("extdata", package = "MOFAcellulaR")
load(file.path(inputs_dir, "testpbcounts.rda"))
load(file.path(inputs_dir, "testcoldata.rda"))

pb_obj <- create_init_exp(counts = testpbcounts,
                          coldata = testcoldata)

ct_list <- filt_profiles(pb_dat = pb_obj,
                         cts = c("Fib","CM"),
                         ncells = 5,
                         counts_col = "cell_counts",
                         ct_col = "cell_type")

ct_list <- filt_views_bysamples(pb_dat_list = ct_list,
                              nsamples = 2)

}
