% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_associations.R
\name{get_associations}
\alias{get_associations}
\title{Associate factors to covariates of interest}
\usage{
get_associations(
  model,
  metadata,
  sample_id_column,
  test_variable,
  test_type = "categorical",
  categorical_type = "parametric",
  group = FALSE
)
}
\arguments{
\item{model}{A MOFA2 model.}

\item{metadata}{A data frame containing the annotations of the samples included in the MOFA model.}

\item{sample_id_column}{A string character that refers to the column in \code{metadata} where the sample identifier is located.}

\item{test_variable}{A string character that refers to the column in \code{metadata} where the covariate to be tested is located.}

\item{test_type}{A string character ("categorical", "continuous").}

\item{categorical_type}{A string character ("parametric", "not_parametric"), only applies for categorical data.}

\item{group}{Boolean flag TRUE/FALSE, to specify if a grouped MOFA model is provided.}
}
\value{
A dataframe in a tidy format containing the p-values of the association tests per factor
}
\description{
Performs Analysis of Variance (ANOVA) or linear models to associate
factor scores with covariates of interest provided by the user
}
\details{
Given a covariate of interest and a defined test, this function tests for associations with factor scores. For
categorical tests, ANOVAs are fitted, while for continous variables, linear models. P-values are corrected
using the Benjamini-Hochberg procedure.
}
\examples{
inputs_dir <- base::system.file("extdata", package = "MOFAcellulaR")
model <- MOFA2::load_model(file.path(inputs_dir, "testmodel.hdf5"))
metadata <- readRDS(file.path(inputs_dir, "testmetadata.rds"))
metadata$var <- stats::rnorm(nrow(metadata))

categorical_assoc <- get_associations(model = model,
                                      metadata = metadata,
                                      sample_id_column = "sample",
                                      test_variable = "patient_group",
                                      test_type = "categorical",
                                      group = FALSE)

continuous_assoc <- get_associations(model = model,
                                      metadata = metadata,
                                      sample_id_column = "sample",
                                      test_variable = "var",
                                      test_type = "continuous",
                                      group = FALSE)
}
