% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tidy_factors.R
\name{get_geneweights}
\alias{get_geneweights}
\title{Extract feature weights scores from a model in a tidy format for a factor of interest}
\usage{
get_geneweights(model, factor)
}
\arguments{
\item{model}{A MOFA2 model.}

\item{factor}{A string character to define the factor to be extracted.}
}
\value{
A dataframe in a tidy format containing the factor feature weights
}
\description{
Generates a tidy dataframe of feature weights for factors of interest useful for plotting function and enrichment analysis
}
\details{
This function simplifies the extraction of factor feature weights from MOFA models.
}
\examples{
inputs_dir <- base::system.file("extdata", package = "MOFAcellulaR")
model <- MOFA2::load_model(file.path(inputs_dir, "testmodel.hdf5"))
gene_weights <- get_geneweights(model = model,
                                factor = "Factor1")
}
