% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tidy_factors.R
\name{get_tidy_factors}
\alias{get_tidy_factors}
\title{Extract factor scores from a model in a tidy format with meta data}
\usage{
get_tidy_factors(
  model,
  metadata,
  factor,
  group = FALSE,
  sample_id_column = "sample"
)
}
\arguments{
\item{model}{A MOFA2 model.}

\item{metadata}{A data frame containing the annotations of the samples included in the MOFA model.}

\item{factor}{A string character to define the factor to be extracted. Alternatively, "all" to get all factors.}

\item{group}{Boolean flag TRUE/FALSE, to specify if a grouped MOFA model is provided.}

\item{sample_id_column}{A string character that refers to the column in \code{metadata} where the sample identifier is located.}
}
\value{
A dataframe in a tidy format containing the factor scores per sample together with metadata
}
\description{
Generates a tidy dataframe for factors of interest useful for plotting functions.
}
\details{
This function simplifies the extraction of factor scores from MOFA models.
}
\examples{
inputs_dir <- base::system.file("extdata", package = "MOFAcellulaR")
model <- MOFA2::load_model(file.path(inputs_dir, "testmodel.hdf5"))
metadata <- readRDS(file.path(inputs_dir, "testmetadata.rds"))
all_factors <- get_tidy_factors(model = model,
                                metadata = metadata,
                                factor = "all",
                                sample_id_column = "sample")
Factor3 <- get_tidy_factors(model = model,
                            metadata = metadata,
                            factor = "Factor3",
                            sample_id_column = "sample")
}
