% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MOFA_hmap.R
\name{plot_MOFA_hmap}
\alias{plot_MOFA_hmap}
\title{Visualize MOFA multicellular model}
\usage{
plot_MOFA_hmap(
  model,
  group = FALSE,
  metadata,
  sample_id_column = "sample",
  sample_anns,
  assoc_list = NULL
)
}
\arguments{
\item{model}{A MOFA2 model.}

\item{group}{Boolean flag TRUE/FALSE, to specify if a grouped MOFA model is provided.}

\item{metadata}{A data frame containing the annotations of the samples included in the MOFA model.}

\item{sample_id_column}{A string character that refers to the column in \code{metadata} where the sample identifier is located.}

\item{sample_anns}{A vector containing strings that refer to the columns in \code{metadata} to be used to annotate samples}

\item{assoc_list}{A named list collecting results of \code{MOFAcellulaR::get_associations()}}
}
\value{
A dataframe in a tidy format containing the manifold and the scatter plot
}
\description{
Plots a heatmap with factor scores annotated by sample's information and factor's statistics
}
\details{
This function summarizes a \code{MOFA2} model by plotting and clustering the factor scores across samples.
Additionally, it allows you to annotate each sample with categorical or continous variables. Finally,
for each factor, the amount of explained variance captured for each cell type is shown. If provided,
summary of the association statistics and sample variables can be provided
}
\examples{
inputs_dir <- base::system.file("extdata", package = "MOFAcellulaR")
model <- MOFA2::load_model(file.path(inputs_dir, "testmodel.hdf5"))
metadata <- readRDS(file.path(inputs_dir, "testmetadata.rds"))
metadata$var <- stats::rnorm(nrow(metadata))

categorical_assoc <- get_associations(model = model,
                                      metadata = metadata,
                                      sample_id_column = "sample",
                                      test_variable = "patient_group",
                                      test_type = "categorical",
                                      group = FALSE)

continuous_assoc <- get_associations(model = model,
                                     metadata = metadata,
                                     sample_id_column = "sample",
                                     test_variable = "var",
                                     test_type = "continous",
                                     group = FALSE)


assoc_list = list("categorical" = categorical_assoc, "continous" = continuous_assoc)

plot_MOFA_hmap(model = model,
               group = FALSE,
               metadata = metadata,
               sample_id_column = "sample",
               sample_anns = c("patient_group", "batch", "var"),
               assoc_list = assoc_list)
}
