% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_2D.R
\name{plot_sample_2D}
\alias{plot_sample_2D}
\title{Visualize sample variability in a 2D space}
\usage{
plot_sample_2D(
  model,
  group = FALSE,
  method = "UMAP",
  metadata,
  sample_id_column = "sample",
  color_by,
  ...
)
}
\arguments{
\item{model}{A MOFA2 model.}

\item{group}{Boolean flag TRUE/FALSE, to specify if a grouped MOFA model is provided.}

\item{method}{A string specifying if "UMAP" or "MDS" should be performed}

\item{metadata}{A data frame containing the annotations of the samples included in the MOFA model.}

\item{sample_id_column}{A string character that refers to the column in \code{metadata} where the sample identifier is located.}

\item{color_by}{A string character that refers to the column in \code{metadata} where the covariate to be tested is located.}

\item{...}{inherited parameters of \code{uwot::umap()}}
}
\value{
A dataframe in a tidy format containing the manifold and the scatter plot
}
\description{
Performs dimensionality reduction of factor scores for the visualization of sample-level variability
}
\details{
For a \code{MOFA2} model it performs a multidimensional scaling (MDS)
or uniform manifold approximation and projection (UMAP) of the factor
scores. It allows you to color samples based on a covariate available in
your meta-data.
}
\examples{
inputs_dir <- base::system.file("extdata", package = "MOFAcellulaR")
model <- MOFA2::load_model(file.path(inputs_dir, "testmodel.hdf5"))
metadata <- readRDS(file.path(inputs_dir, "testmetadata.rds"))
UMAP_embedding <- plot_sample_2D(model = model,
                                 group = FALSE,
                                 method = "UMAP",
                                 metadata = metadata,
                                 sample_id_column = "sample",
                                 color_by = "batch")
}
