% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{tmm_trns}
\alias{tmm_trns}
\title{TMM normalization of single cell data sets}
\usage{
tmm_trns(pb_dat_list, scale_factor = 1e+06)
}
\arguments{
\item{pb_dat_list}{List of SummarizedExperiment generated from \code{MOFAcellulaR::filt_profiles()}}

\item{scale_factor}{Numeric. Scaled counts are multiplied by this factor before log transformation}
}
\value{
A named list of SummarizedExperiments per cell type provided with normalized log transformed data in their \code{logcounts} assay
}
\description{
Performs for a list of SummarizedExperiments, TMM normalization scaled by a factor specified by the user.
}
\details{
This function estimates TMM normalization factors and normalizes a list of gene count matrices, data
is additionally scaled using a factor specified by the user and \code{log1p()} transformed
}
\examples{
inputs_dir <- base::system.file("extdata", package = "MOFAcellulaR")
load(file.path(inputs_dir, "testpbcounts.rda"))
load(file.path(inputs_dir, "testcoldata.rda"))

pb_obj <- create_init_exp(counts = testpbcounts,
                          coldata = testcoldata)

ct_list <- filt_profiles(pb_dat = pb_obj,
                         cts = c("Fib","CM"),
                         ncells = 5,
                         counts_col = "cell_counts",
                         ct_col = "cell_type")

ct_list <- filt_gex_byexpr(pb_dat_list = ct_list,
                           min.count = 5,
                           min.prop = 0.25)

ct_list <- tmm_trns(pb_dat_list = ct_list,
                    scale_factor = 1000000)
}
